Fixed average score creation.

+++ db/migrate/20111117230849_add_average_score_to_cifs.rb
@@ -1,5 +1,5 @@
 class AddAverageScoreToCifs < ActiveRecord::Migration
   def change
-    add_column :cifs, :average_score, :double
+    add_column :cifs, :average_score, :float
   end
 end
